requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchaudio
flask
Pillow
stability_sdk
rembg
transformers
pyopenssl